Refactor and add navigation schema
