+++ app/build.gradle
@@ -48,6 +48,7 @@ android {
 
 dependencies {
 
+    implementation 'androidx.navigation:navigation-fragment:2.5.3'
     def composeBom = platform('androidx.compose:compose-bom:2023.01.00')
     implementation composeBom
     androidTestImplementation composeBom
