--- app/src/main/java/com/github/mateo762/myapplication/fragments/CalendarFragment.kt
@@ -1,19 +1,0 @@
-package com.github.mateo762.myapplication.fragments
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.github.mateo762.myapplication.R
-
-class CalendarFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_calendar, container, false)
-    }
-}
