+++ app/src/main/java/com/github/mateo762/myapplication/home/HomeActivity.kt
@@ -1,0 +1,108 @@
+package com.github.mateo762.myapplication.home
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import com.firebase.ui.auth.AuthUI
+import com.firebase.ui.auth.FirebaseAuthUIActivityResultContract
+import com.firebase.ui.auth.data.model.FirebaseAuthUIAuthenticationResult
+import com.github.mateo762.myapplication.BaseActivity
+import com.github.mateo762.myapplication.R
+import com.github.mateo762.myapplication.home.fragments.FeedFragment
+import com.github.mateo762.myapplication.home.fragments.SummaryFragment
+import com.github.mateo762.myapplication.home.fragments.TodayFragment
+import com.google.android.gms.auth.api.identity.Identity
+import com.google.android.gms.auth.api.identity.SignInClient
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.google.firebase.auth.FirebaseAuth
+
+class HomeActivity : BaseActivity() {
+
+    private lateinit var bottomNavView: BottomNavigationView
+    private lateinit var oneTapClient: SignInClient
+    private val signInLauncher = registerForActivityResult(
+        FirebaseAuthUIActivityResultContract()
+    ) { res ->
+        this.onSignInResult(res)
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_home)
+        super.onCreateDrawer();
+        bottomNavView = findViewById(R.id.bottomNav)
+        bottomNavView.setOnItemSelectedListener(bottomNavListener)
+
+        oneTapClient = Identity.getSignInClient(this)
+
+    }
+
+    private val bottomNavListener = BottomNavigationView.OnNavigationItemSelectedListener {
+        lateinit var selectedFragment: Fragment
+        when (it.itemId) {
+            R.id.todayFragment -> {
+                selectedFragment = TodayFragment()
+            }
+            R.id.feedFragment -> {
+                selectedFragment = FeedFragment()
+            }
+            R.id.summaryFragment -> {
+                selectedFragment = SummaryFragment()
+            }
+        }
+        openFragmentSelected(selectedFragment)
+        true
+    }
+
+    private fun openFragmentSelected(fragment: Fragment): Int {
+        return supportFragmentManager.beginTransaction().replace(
+            R.id.navHostFragment, fragment
+        ).commit()
+    }
+
+    private fun createSignInIntent() {
+        // Choose authentication providers
+        val providers = arrayListOf(
+            AuthUI.IdpConfig.EmailBuilder().build(),
+            AuthUI.IdpConfig.PhoneBuilder().build(),
+            AuthUI.IdpConfig.GoogleBuilder().build()
+        )
+
+        // Create and launch sign-in intent
+        val signInIntent = AuthUI.getInstance()
+            .createSignInIntentBuilder()
+            .setAvailableProviders(providers)
+            .build()
+        signInLauncher.launch(signInIntent)
+    }
+
+    private fun onSignInResult(result: FirebaseAuthUIAuthenticationResult) {
+        val response = result.idpResponse
+        if (result.resultCode == RESULT_OK) {
+            // Successfully signed in
+            val user = FirebaseAuth.getInstance().currentUser
+            println("Sign in SUCCESSFULLY COMPLETED, user = $user")
+        } else {
+            // Sign in failed. If response is null the user canceled the
+            // sign-in flow using the back button. Otherwise check
+            // and handle the error.
+            val code = response!!.error!!.errorCode
+            println("Code = $code")
+        }
+    }
+
+    private fun signOut() {
+        AuthUI.getInstance()
+            .signOut(this)
+            .addOnCompleteListener {
+                // ...
+            }
+    }
+
+    private fun delete() {
+        AuthUI.getInstance()
+            .delete(this)
+            .addOnCompleteListener {
+                // ...
+            }
+    }
+}
