+++ app/src/main/res/menu/habits_menu_bottom.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/developFragment"
+        android:icon="@drawable/blur_icon"
+        android:title="Develop"/>
+
+    <item
+        android:id="@+id/weekFragment"
+        android:icon="@drawable/blur_icon"
+        android:title="Week"/>
+
+    <item
+        android:id="@+id/listFragment"
+        android:icon="@drawable/blur_icon"
+        android:title="List"/>
+
+</menu>
