+++ app/src/main/res/menu/menu_bottom.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/todayFragment"
+        android:icon="@drawable/blur_icon"
+        android:title="Today"/>
+
+    <item
+        android:id="@+id/feedFragment"
+        android:icon="@drawable/blur_icon"
+        android:title="Feed"/>
+
+    <item
+        android:id="@+id/summaryFragment"
+        android:icon="@drawable/blur_icon"
+        android:title="Summary"/>
+
+</menu>
