+++ app/src/main/res/menu/nav_menu.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:showIn="navigation_view">
+    <group android:checkableBehavior="single">
+        <item
+            android:id="@+id/nav_home"
+            android:icon="@drawable/blur_icon"
+            android:title="Home" />
+        <item
+            android:id="@+id/nav_habits"
+            android:icon="@drawable/blur_icon"
+            android:title="Habits" />
+        <item
+            android:id="@+id/nav_search"
+            android:icon="@drawable/blur_icon"
+            android:title="Search" />
+        <item
+            android:id="@+id/nav_profile"
+            android:icon="@drawable/blur_icon"
+            android:title="Profile" />
+    </group>
+
+    <item android:title="">
+        <menu>
+            <item
+                android:id="@+id/nav_settings"
+                android:icon="@drawable/blur_icon"
+                android:title="Settings" />
+            <item
+                android:id="@+id/nav_logout"
+                android:icon="@drawable/blur_icon"
+                android:title="Log Out" />
+        </menu>
+    </item>
+</menu>
