+++ app/src/main/res/navigation/habits_nav_graph.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/habits_nav_graph"
+    app:startDestination="@id/developFragment">
+
+    <fragment
+        android:id="@+id/developFragment"
+        android:name="com.github.mateo762.myapplication.habits.fragments.DevelopFragment"
+        android:label="fragment_develop"
+        tools:layout="@layout/fragment_develop" />
+    <fragment
+        android:id="@+id/weekFragment"
+        android:name="com.github.mateo762.myapplication.habits.fragments.WeekFragment"
+        android:label="fragment_week"
+        tools:layout="@layout/fragment_week" />
+    <fragment
+        android:id="@+id/listFragment"
+        android:name="com.github.mateo762.myapplication.habits.fragments.ListFragment"
+        android:label="fragment_list"
+        tools:layout="@layout/fragment_list" />
+</navigation>
